disabled zoom in button
zooming by drawing a box and zooming to that area was causing issues on
some systems where the zoom would not wait until the box was finished
drawing (left mouse button up).
Zooming in this way is not actually required because we already have a
way of disambiguating between overlapping vowels
Note: the code still is in place. The lines involved in drawing the
zoom button have merely been commented out.